chore: Reduce grafana status update requests. (#2057)
* chore(NamespacedResource): Remove unused functions

* chore(shared_controller): Helpers for adding/removing a namespaced resource to a Grafana Status list

* chore(Types): Add/update NamespacedResource func

* chore(Controllers): Refactor to use new helper functions

* chore(NamespacedResourceList): Delete obsolete Add function

* chore(Grafana): Move Grafana status mutations under Grafana struct

* test(NamespacedResource): add/remove from status lists

* test: Ensure all namespaced resources are tested against Grafana Status

* test(types): Ensure Types are implemented except Notification Policy and Route

* fix(NamespacedResource): Jsonpatch op was add, not remove

* chore(NamespacedResource): Simplify removal to a single case


* chore(LibraryPanel): Remove ResyncHasElapsedCheck

* chore(LibraryPanel): Update Grafana CR after Create requests

* chore(NamespacedResourceList): Refactor Find, IndexOf, and Delete to use slices.IndexFunc


* test: NamespacedResource helper functions

* fix(Controllers): Ensure GrafanaStatus is always updated

* chore: Apply suggestions


---------

+++ controllers/alertrulegroup_controller.go
@@ -255,16 +255,17 @@ func (r *GrafanaAlertRuleGroupReconciler) reconcileWithInstance(ctx context.Cont
 	}
 
 	// Update grafana instance Status
-	instance.Status.AlertRuleGroups = instance.Status.AlertRuleGroups.Add(group.Namespace, group.Name, mGroup.Title)
-	return r.Client.Status().Update(ctx, instance)
+	return instance.AddNamespacedResource(ctx, r.Client, group, group.NamespacedResource())
 }
 
 func (r *GrafanaAlertRuleGroupReconciler) finalize(ctx context.Context, group *grafanav1beta1.GrafanaAlertRuleGroup) error {
 	log := logf.FromContext(ctx)
+
+	isCleanupInGrafanaRequired := true
 	folderUID, err := getFolderUID(ctx, r.Client, group)
 	if err != nil {
-		log.Info("ignoring finalization logic as folder no longer exists")
-		return nil //nolint:nilerr
+		log.Info("Skipping Grafana finalize logic as folder no longer exists")
+		isCleanupInGrafanaRequired = false
 	}
 
 	instances, err := GetScopedMatchingInstances(ctx, r.Client, group)
@@ -273,22 +274,25 @@ func (r *GrafanaAlertRuleGroupReconciler) finalize(ctx context.Context, group *g
 	}
 
 	for _, instance := range instances {
-		cl, err := client2.NewGeneratedGrafanaClient(ctx, r.Client, &instance)
-		if err != nil {
-			return fmt.Errorf("building grafana client: %w", err)
-		}
+		// Skip cleanup in instances
+		if isCleanupInGrafanaRequired {
+			cl, err := client2.NewGeneratedGrafanaClient(ctx, r.Client, &instance)
+			if err != nil {
+				return fmt.Errorf("building grafana client: %w", err)
+			}
 
-		_, err = cl.Provisioning.DeleteAlertRuleGroup(group.GroupName(), folderUID) //nolint:errcheck
-		if err != nil {
-			var notFound *provisioning.DeleteAlertRuleGroupNotFound
-			if !errors.As(err, &notFound) {
-				return fmt.Errorf("deleting alert rule group: %w", err)
+			_, err = cl.Provisioning.DeleteAlertRuleGroup(group.GroupName(), folderUID) //nolint:errcheck
+			if err != nil {
+				var notFound *provisioning.DeleteAlertRuleGroupNotFound
+				if !errors.As(err, &notFound) {
+					return fmt.Errorf("deleting alert rule group: %w", err)
+				}
 			}
 		}
 
 		// Update grafana instance Status
-		instance.Status.AlertRuleGroups = instance.Status.AlertRuleGroups.Remove(group.Namespace, group.Name)
-		if err = r.Client.Status().Update(ctx, &instance); err != nil {
+		err = instance.RemoveNamespacedResource(ctx, r.Client, group)
+		if err != nil {
 			return err
 		}
 	}
